Move shaping logic into selector
