+++ web/src/features/fireWatch/slices/burnForecastSlice.ts
@@ -1,7 +1,9 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit'
+import { createSelector, createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { AppThunk } from 'app/store'
 import { getBurnForecasts } from '@/features/fireWatch/fireWatchApi'
-import { FireWatchBurnForecast } from '@/features/fireWatch/interfaces'
+import { BurnForecast, FireWatchBurnForecast, PrescriptionEnum } from '@/features/fireWatch/interfaces'
+import { RootState } from '@/app/rootReducer'
+import { BurnWatchRow } from '@/features/fireWatch/components/FireWatchDashboard'
 
 export interface BurnForecastsState {
   loading: boolean
@@ -53,3 +55,39 @@ export const fetchBurnForecasts = (): AppThunk =>
       dispatch(getBurnForecastsFailed((err as Error).toString()))
     }
   }
+
+const selectFireWatchBurnForecasts = (state: RootState) => state.burnForecasts.fireWatchBurnForecasts
+
+export const selectBurnForecasts = createSelector([selectFireWatchBurnForecasts], burnForecasts => {
+  const newBurnForecasts: BurnWatchRow[] =  burnForecasts.map((value: FireWatchBurnForecast) => {
+    return {
+      id: value.fireWatch.id,
+      title: value.fireWatch.title,
+      fireCentre: value.fireWatch.fireCentre?.name ?? '',
+      station: value.fireWatch.station?.name ?? '',
+      fuelType: value.fireWatch.fuelType,
+      status: value.fireWatch.status,
+      burnWindowStart: value.fireWatch.burnWindowStart,
+      burnWindowEnd: value.fireWatch.burnWindowEnd,
+      inPrescription: getInPrescription(value.burnForecasts),
+      fireWatch: value.fireWatch,
+      burnForecasts: value.burnForecasts
+    }
+  })
+  return newBurnForecasts
+})
+
+  const getInPrescription = (burnForecasts: BurnForecast[]): PrescriptionEnum => {
+    let inPrescription = PrescriptionEnum.NO
+    for (const burnForecast of burnForecasts) {
+      if (burnForecast.inPrescription === PrescriptionEnum.HFI) {
+        inPrescription = PrescriptionEnum.HFI
+      }
+      if (burnForecast.inPrescription === PrescriptionEnum.ALL) {
+        inPrescription = PrescriptionEnum.ALL
+        break
+      }
+    }
+    return inPrescription
+  }
+
